Backend/requirements.txt
djangorestframework
django
scipy